+++ src/posts/add-post/add-post.component.ts
@@ -52,7 +52,7 @@ export class AddPostComponent implements OnInit {
       } else if(file.type.indexOf('video')> -1){
         this.format = 'video';
       }
-      
+      this.current_post_data.fileType = this.format
       reader.onload = (event) => { // called once readAsDataURL is completed
         this.url = event.target!.result
         
@@ -64,8 +64,13 @@ export class AddPostComponent implements OnInit {
     }
   }
 
-  upload_post() {
-    this.post_service.Upload_post(this.current_post_data)
+  upload_post(caption:any) {
+    this.current_post_data.caption = caption
+    console.log("current_post_data before upload:", this.current_post_data)
+    this.current_post_data = this.post_service.Upload_post(this.current_post_data)
+    console.log("current_post_data after upload:", this.current_post_data)
+    this.post_service.save_metadata_of_post(this.current_post_data)
+    console.log("current_post_data after updating firestore:", this.current_post_data)
   }